requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.28.2
drf-spectacular==0.26.2
model-bakery==1.11.0
